*: factor out resuable parts from store pkg
