+++ cmd/oklog/store.go
@@ -17,6 +17,7 @@ import (
 	"github.com/rs/cors"
 
 	"github.com/oklog/oklog/pkg/cluster"
+	"github.com/oklog/oklog/pkg/event"
 	"github.com/oklog/oklog/pkg/fs"
 	"github.com/oklog/oklog/pkg/group"
 	"github.com/oklog/oklog/pkg/store"
@@ -218,7 +219,7 @@ func runStore(args []string) error {
 		fsys,
 		stagingPath,
 		*segmentTargetSize, *segmentBufferSize,
-		store.LogReporter{Logger: log.With(logger, "component", "FileLog")},
+		event.LogReporter{Logger: log.With(logger, "component", "FileLog")},
 	)
 	if err != nil {
 		return err
@@ -287,7 +288,7 @@ func runStore(args []string) error {
 			consumedBytes,
 			replicatedSegments.WithLabelValues("egress"),
 			replicatedBytes.WithLabelValues("egress"),
-			store.LogReporter{Logger: log.With(logger, "component", "Consumer")},
+			event.LogReporter{Logger: log.With(logger, "component", "Consumer")},
 		)
 		g.Add(func() error {
 			c.Run()
@@ -303,7 +304,7 @@ func runStore(args []string) error {
 			topicPath,
 			*segmentTargetSize,
 			*segmentBufferSize,
-			store.LogReporter{Logger: log.With(logger, "component", "TopicLogs")},
+			event.LogReporter{Logger: log.With(logger, "component", "TopicLogs")},
 		)
 	}
 	{
@@ -312,7 +313,7 @@ func runStore(args []string) error {
 			return err
 		}
 		// TODO(fabxc): track delay between staging and demux completion in a metric.
-		reporter := store.LogReporter{Logger: log.With(logger, "component", "Demuxer")}
+		reporter := event.LogReporter{Logger: log.With(logger, "component", "Demuxer")}
 		demux := store.NewDemuxer(stagingLog, topicLogs, reporter)
 
 		g.Add(func() error {
@@ -332,14 +333,14 @@ func runStore(args []string) error {
 				compactDuration,
 				trashedSegments,
 				purgedSegments,
-				store.LogReporter{Logger: log.With(logger, "component", "Compacter", "topic", t)},
+				event.LogReporter{Logger: log.With(logger, "component", "Compacter", "topic", t)},
 			)
 		}
 		topicLogs, err := newTopicLogs()
 		if err != nil {
 			return err
 		}
-		reporter := store.LogReporter{Logger: log.With(logger, "component", "TopicCompacter")}
+		reporter := event.LogReporter{Logger: log.With(logger, "component", "TopicCompacter")}
 		compacters := store.NewTopicCompacters(topicLogs, cfac, reporter)
 
 		g.Add(func() error {
@@ -365,7 +366,7 @@ func runStore(args []string) error {
 				replicatedSegments.WithLabelValues("ingress"),
 				replicatedBytes.WithLabelValues("ingress"),
 				apiDuration,
-				store.LogReporter{Logger: log.With(logger, "component", "API")},
+				event.LogReporter{Logger: log.With(logger, "component", "API")},
 			)
 			defer func() {
 				if err := api.Close(); err != nil {
